+++ otsdaq-mu2e/ROCCore/ROCCoreVInterface.h
@@ -67,15 +67,15 @@ class ROCCoreVInterface : public FEVInterface
 	//		writeROCRegister/writeEmulatorRegister 
 	//       ... let's wait for the moment to implement writeBlocks
 	void         							writeRegister				(uint16_t address, uint16_t writeData);  // chooses ROC or Emulator version
-	int          							readRegister				(uint16_t address);     // chooses ROC or Emulator version
+	uint16_t        						readRegister				(uint16_t address);     // chooses ROC or Emulator version
 	void 									readBlock					(std::vector<uint16_t>& data, uint16_t address, uint16_t wordCount, bool incrementAddress);     // chooses ROC or Emulator version
 
 	virtual void 							writeROCRegister			(uint16_t address, uint16_t writeData) = 0;  // pure virtual, must define in inheriting children
-	virtual int 							readROCRegister				(uint16_t address) = 0;  // pure virtual, must define in inheriting children
+	virtual uint16_t						readROCRegister				(uint16_t address) = 0;  // pure virtual, must define in inheriting children
 	virtual void 							readROCBlock				(std::vector<uint16_t>& data, uint16_t address, uint16_t wordCount, bool incrementAddress) = 0; // pure virtual, must define in inheriting children
 
 	virtual void 							writeEmulatorRegister		(uint16_t address, uint16_t writeData) = 0;  // pure virtual, must define in inheriting children
-	virtual int 							readEmulatorRegister		(uint16_t address) = 0;  // pure virtual, must define in inheriting children
+	virtual uint16_t 						readEmulatorRegister		(uint16_t address) = 0;  // pure virtual, must define in inheriting children
 	virtual void 							readEmulatorBlock			(std::vector<uint16_t>& data, uint16_t address, uint16_t wordCount, bool incrementAddress) = 0; // pure virtual, must define in inheriting children
